+++ docs/ChangeLog.md
@@ -105,6 +105,7 @@
   - util: suppress false warnings of `bind` inside non-interactive shells (reported by wukuan405) `#D1823` 1e19a67
 - history: support `bleopt history_erasedups_limit` (motivated by SuperSandro2000) `#D1822` e4afb5a 3110967
 - prompt: support `bleopt prompt_{emacs,vi}_mode_indicator` (motivated by ferdinandyb) `#D1843` 2b905f8
+- util (`ble-import`): support option `-q` `#D1859` XXXXXXX
 
 ## Changes